refactor: alertCarInputError 함수에서 if-else if문 삭제
